create useSteppedValues hook
